feat: Player 내 로직 Member 로 이동
